refactor(clone): nest clone/pull/push behind mcx vfs subcommand
Consolidates three top-level commands into `mcx vfs`:
  mcx vfs clone confluence <space> [dir]
  mcx vfs pull [dir]
  mcx vfs push [dir]

Keeps the top-level command surface clean as requested.
Removes standalone clone.ts, pull.ts, push.ts command files.
